% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handling.R
\name{udm_error_body}
\alias{udm_error_body}
\title{Helper function to return error message to 'udmpullr'}
\usage{
udm_error_body(resp)
}
\arguments{
\item{resp}{A httr2 response object}
}
\value{
A message for use in \code{httr2::req_error}
}
\description{
Helper function to return error message to 'udmpullr'
}
