% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmpullr-package.R
\docType{package}
\name{udmpullr-package}
\alias{udmpullr-package}
\title{udmpullr: Access Data from the 'Understanding Data and Markets' API}
\description{
Provides functions to access data from the 'Understanding Data and Markets' (UDM) API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmhanon/udmpullr}
  \item Report bugs at \url{https://github.com/tmhanon/udmpullr/issues}
}

}
\author{
\strong{Maintainer}: Tristan Hanon \email{tristan.hanon@gmail.com} (\href{https://orcid.org/0009-0000-3684-9717}{ORCID})

}
\keyword{internal}
