% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udmpullr.R
\name{udmpullr}
\alias{udmpullr}
\title{Pull tables from 'UDM' API}
\usage{
udmpullr(
  table_name,
  product_name = NULL,
  start_date = NULL,
  end_date = NULL,
  company_feed = "Universal"
)
}
\arguments{
\item{table_name}{Name of the table you want to pull}

\item{product_name}{Name of a specific product to filter for (if Product_Name is a variable in the table)}

\item{start_date}{Earliest date to include in the dataset}

\item{end_date}{Latest date to include in the dataset}

\item{company_feed}{Your company feed for 'UDM' (if applicable). Default is "Universal"}
}
\value{
A tibble
}
\description{
Use this function to pull tables from the 'UDM' API.
}
\examples{
\dontrun{
udmpullr(table_name = "UDM_Report_CME_Dairy_Spot_Market_Overview_Daily")
}
}
