% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udm_credentials.R
\name{udm_credentials}
\alias{udm_credentials}
\title{Set 'UDM' API credentials}
\usage{
udm_credentials(UDM_API_USERNAME, UDM_API_PASSWORD)
}
\arguments{
\item{UDM_API_USERNAME}{Your email address registered with UDM}

\item{UDM_API_PASSWORD}{Your UDM password}
}
\description{
In order to access the 'UDM' API you must have an account registered with 'UDM'.
Visit (https://commoditymarkets.com/) for more information.
}
\details{
This function provides one option for telling R your UDM credentials to use
within your current session. If you use the 'UDM' API regularly, a better option
is to use environmental variables to permanently set your credentials. Open or
create a .Renviron file using \code{usethis::edit_r_environ} and add
UDM_API_USERNAME="<your UDM email/username>" and UDM_API_PASSWORD="\if{html}{\out{<your UDM password>}}"
on separate lines. Save the file (ensuring there is a blank line at the end of the
file) and restart R. Your 'UDM' API credentials will now be available to \code{udmpullr}
across R sessions.
}
\examples{
\dontrun{
# Set your UDM credentials
udm_credentials("<your UDM email/username>", "<your UDM password")

# Retrieve UDM credentials
Sys.getenv("UDM_API_USERNAME")
Sys.getenv("UDM_API_PASSWORD")
}
}
